requirements/dev.txt
pytest==7.*
pytest-django==4.*
